+++ front/components/assistant/conversation/hooks/suggestion.ts
@@ -1,7 +1,7 @@
 import { ReactRenderer } from "@tiptap/react";
 import tippy from "tippy.js";
 
-import { MentionList } from "./MentionList";
+import { MentionList } from "@app/components/assistant/conversation/hooks/MentionList";
 
 export interface EditorSuggestion {
   id: string;
