+++ front/components/assistant/conversation/hooks/useAssistantSuggestions.ts
